Authent using Micronaut LoginController
