+++ src/main/java/com/michelin/ns4kafka/security/gitlab/package-info.java
@@ -1,0 +1,8 @@
+@Configuration
+@Requires(property = "micronaut.security.gitlab.enabled", notEquals = StringUtils.FALSE)
+package com.michelin.ns4kafka.security.gitlab;
+
+import io.micronaut.context.annotation.Configuration;
+import io.micronaut.context.annotation.Requires;
+import io.micronaut.core.util.StringUtils;
+
